+++ Completed/WiriScript v4_CN.lua
